refactor(knowledge): 优化检索功能实现并支持段落级别排除
- 引入段落服务接口和实体类支持段落级别操作
- 实现段落活跃状态过滤，排除非活跃段落参与检索
- 修复文档服务中变量命名问题(noActiveDocuments替换excludeDocuments)
- 统一分数类型从float改为double提升精度
- 重构向量存储实现在搜索时进行去重并按分数排序
- 更新嵌入存储代理配置移除注释信息
- 修改搜索请求参数移除分页相关字段简化接口
- 修复问题段落服务中段落ID映射错误
- 优化复合存储和全文存储中的分数处理逻辑
- 添加段落级别排除ID列表支持更精确的检索控制

+++ maxkb4j-service-api/maxkb4j-knowledge-api/src/main/java/com/maxkb4j/knowledge/entity/EmbeddingEntity.java
@@ -44,7 +44,7 @@ public class EmbeddingEntity {
 	private String content;
 	@TableField(exist = false)
 	@TextScore
-	private float score; // 匹配度得分
+	private double score; // 匹配度得分
 	@Transient
 	private Integer dimension;
